requirements.txt
firecrawl-py
qdrant-client
streamlit
fastembed
firecrawl
TTS
python-dotenv 
openai-agents
python-dotenv
gTTS
sentence_transformers